middleware fixes

+++ routes/web.php
@@ -26,8 +26,8 @@ Route::get('/contact', 'ContactFormControler@create');
 Route::post('/contact', 'ContactFormControler@store');
 
 //Route::post('/contact', 'homeController@store');
-Route::get('/perfil', 'homeController@perfil');
-Route::get('/perfilAdm', 'homeController@perfilAdm')->middleware('Admin');
+Route::get('/perfil', 'homeController@perfil') ->middleware('auth');
+Route::get('/perfilAdm', 'homeController@perfilAdm')->middleware('admin');
 //Route::get('/login', 'homeController@login');
 // Route::get('/registro', 'homeController@register');
 
@@ -38,7 +38,7 @@ Route::get('product-detail/{id}', 'ProductController@detail');
 Route::get('/products/create', 'ProductController@create')->name('products.create');
 Route::post('/products/create', 'ProductController@store');
 Route::get('/products/show','ProductController@show');
-Route::get('/products/{id}/update','ProductController@edit')->name('products.edit')->middleware('Admin');
+Route::get('/products/{id}/update','ProductController@edit')->name('products.edit')->middleware('admin');
 Route::patch('/products/{id}/update', 'ProductController@update')->name('products.update');
 Route::delete('/products/{id}/', 'ProductController@destroy')->name('products.destroy');
 Route::post('products/search', 'ProductController@search')->name('products.search');
@@ -50,10 +50,10 @@ Route::get('/deleteCart/{id}', 'CartController@deleteCart');
 Route::get('/checkout', 'CartController@checkout');
 // USERS|
 Route::get('/users/index', 'UsersController@index')->name('users.index');
-Route::get('/users/show/{id}', 'UsersController@show')->name('users.show');
-Route::get('/users/{id}/edit/', 'UserController@edit')->name('user.edit');
-Route::post('/users/update/{id}', 'UserController@update');
-Route::delete('users/{id}', 'UserController@destroy')->name('users.detroy');
+Route::get('/users/show/{id}', 'UsersController@show')->name('users.show')->middleware('auth');
+Route::get('/users/{id}/edit/', 'UserController@edit')->name('user.edit')->middleware('auth');
+Route::post('/users/update/{id}', 'UserController@update')->middleware('auth');
+Route::delete('users/{id}', 'UserController@destroy')->name('users.detroy')->middleware('auth');
